refactor(execution): extract single-transition runner into TransitionExecutor (S2)
Move the one-transition step loop (RunSingleTransitionAsync and its helpers:
BuildExecutionPlan, CreateInitialState, DetermineFlowControl, NeedsReplan,
EnrichTelemetry, BuildLogScope, ExecuteStepWithBoundary, PipelineState/
FlowControl) out of TransitionPipeline into a dedicated TransitionExecutor.

TransitionExecutor.ExecuteOneAsync owns no lock, no auto-chain loop and no
enqueue; it runs the ordered steps for one already-validated/profiled context
applying StepOutcome flow control. TransitionPipeline keeps lock acquisition,
the auto-chain loop, post-commit orchestration and fault/resolve handling, and
now delegates per-transition execution to the executor.

Behavior-preserving: identical step logic, same namespace, [Trace] retained.
TransitionPipeline ctor now takes TransitionExecutor instead of the step
enumerable; DI and TransitionPipelineTests updated accordingly.
